add restart-all

+++ stop.sh
@@ -4,7 +4,7 @@ if [[ "$1" == "pero" ]] && [ -f ./pero_pid.txt ]; then
     kill `cat pero_pid.txt`
     kill -9 `cat pero_pid.txt`
     rm pero_pid.txt
-elif [[ "$1" == "meows" ]]; then
+elif [[ "$1" == "meows" ]] && [ -f ./meows_pid.txt ]; then
     kill `cat meows_pid.txt`
     kill -9 `cat meows_pid.txt`
     rm meows_pid.txt
